agent_tools/rag_mcp_server/requirements.txt
python-docx>=1.2.0
mcp
python-dotenv
google-cloud-aiplatform